FAKEai-service-v2/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.10.6
pydantic-settings==2.1.0

# PDF extraction
PyMuPDF==1.23.7

# NumPy
numpy>=1.21,<2.0

# OCR
pytesseract==0.3.10
Pillow==10.1.0
opencv-contrib-python==4.10.0.84
paddlepaddle==3.2.2
paddleocr==3.2.0

# Tax ID validation
python-stdnum==1.20

# HTTP client (AI fallback)
httpx>=0.28.1,<1.0.0

# Mistral OCR
mistralai>=1.5.0,<2.0.0

# Testing
pytest==8.3.5
pytest-asyncio==0.25.0
